+++ src/sankeyDiagram.ts
@@ -98,6 +98,8 @@ import {
     FontSettingsOptions,
     ButtonSettings,
     buttonDefaults,
+    LinkMatchType,
+    LinkColorContainerItem,
 } from "./settings";
 import { FormattingSettingsService } from "powerbi-visuals-utils-formattingmodel";
 
@@ -138,12 +140,16 @@ export class SankeyDiagram implements IVisual {
     private static resetButton: ClassAndSelector = createClassAndSelector("resetButton");
     private static ResetButtonRect: ClassAndSelector = createClassAndSelector("resetButtonRect");
     private static ResetButtonText: ClassAndSelector = createClassAndSelector("resetButtonText");
-
-    private static DefaultColourOfLink: string = "#4F4F4F";
+    private static StrokeVisibleClass: ClassAndSelector = createClassAndSelector("strokeVisible");
 
     private static LinksPropertyIdentifier: DataViewObjectPropertyIdentifier = {
         objectName: "links",
         propertyName: "fill"
+    };
+
+    private static LinksBorderPropertyIdentifier: DataViewObjectPropertyIdentifier = {
+        objectName: "links",
+        propertyName: "borderColor"
     };
 
     private static NodesPropertyIdentifier: DataViewObjectPropertyIdentifier = {
@@ -451,7 +457,7 @@ export class SankeyDiagram implements IVisual {
         const name: string = node.levelValues?.[0]?.value?.toString() ?? null;
         const nodeFillColor = this.getColor(
             SankeyDiagram.NodesPropertyIdentifier,
-            this.colorPalette.getColor(name).value,
+            settings.nodes.defaultContainerItem.fill.value.value ?? this.colorPalette.getColor(name).value,
             node.objects);
         const nodeStrokeColor = this.colorHelper.getHighContrastColor("foreground", nodeFillColor);
 
@@ -464,7 +470,7 @@ export class SankeyDiagram implements IVisual {
             outputWeight: 0,
             backwardWeight: 0,
             selfLinkWeight: 0,
-            width: settings.nodesSettings.nodeWidth.value,
+            width: settings.nodes.defaultContainerItem.nodeWidth.value,
             height: 0,
             columnIndex: SankeyDiagram.DefaultIndex,
             fillColor: nodeFillColor,
@@ -559,11 +565,15 @@ export class SankeyDiagram implements IVisual {
                         Number(child.values[weightIndex].value) || SankeyDiagram.DefaultWeightValue : SankeyDiagram.MinWeightValue;
                     weightValues.push(weight);
                 }
-                const linkFillColor = this.getColor(
+
+                const linkFillColor = this.getLinkColor(
+                    settings,
                     SankeyDiagram.LinksPropertyIdentifier,
-                    SankeyDiagram.DefaultColourOfLink,
-                    child.objects);
-                const linkStrokeColor = this.colorHelper.isHighContrast ? this.colorHelper.getHighContrastColor("foreground", linkFillColor) : linkFillColor;
+                    settings.links.defaultContainerItem.fill.value.value || LinkColorContainerItem.DefaultColorOfLink,
+                    child.objects,
+                    foundSource,
+                    foundDestination);
+                const linkStrokeColor = this.colorHelper.getHighContrastColor("foreground", this.getLinkBorderColor(SankeyDiagram.LinksBorderPropertyIdentifier, linkFillColor, child.objects, settings));
 
                 const valuesFormatterForLinkTooltipInfo = valueFormatter.create({
                     format: formatOfWeight,
@@ -637,6 +647,7 @@ export class SankeyDiagram implements IVisual {
         // add ColorPicker for each node and link to the Format pane
         this.sankeyDiagramSettings.populateNodesColorSelector(nodes);
         this.sankeyDiagramSettings.populateLinksColorSelector(links);
+        this.sankeyDiagramSettings.handleHighContrastMode(this.colorHelper);
 
         const sankeyDiagramDataView = {
             nodes,
@@ -926,6 +937,34 @@ export class SankeyDiagram implements IVisual {
         return setting;
     }
 
+
+    private getLinkColor(
+        settings: SankeyDiagramSettings,
+        properties: DataViewObjectPropertyIdentifier,
+        defaultColor: string,
+        objects: DataViewObjects,
+        source: SankeyDiagramNode,
+        destination: SankeyDiagramNode): string {
+
+        const color: string = this.getColor(properties, defaultColor, objects);
+        if (settings.links.matchNodeColors.value) {
+            switch (settings.links.matchSourceOrDestination.value.value) {
+                case LinkMatchType.Source:
+                    return source.fillColor || color;
+                case LinkMatchType.Destination:
+                    return destination.fillColor || color;
+            }
+        }
+        return color;
+    }
+
+    private getLinkBorderColor(properties: DataViewObjectPropertyIdentifier, linkColor: string, objects: DataViewObjects, settings: SankeyDiagramSettings): string{
+        const defaultColor = settings.links.defaultContainerItem.border.color.value.value || linkColor;
+        const color: string = this.getColor(properties, defaultColor, objects);
+
+        return this.colorHelper.getHighContrastColor("foreground", color);
+    }
+
     private getColor(
         properties: DataViewObjectPropertyIdentifier,
         defaultColor: string,
@@ -1077,9 +1116,6 @@ export class SankeyDiagram implements IVisual {
             settings.sort = foundSortedColumn.displayName + "|" + foundSortedColumn.sort;
         }
 
-        // change settings from high contrast mode
-        settings.labels.fill.value.value = this.colorHelper.getHighContrastColor("foreground", settings.labels.fill.value.value);
-        settings.linkLabels.fill.value.value = this.colorHelper.getHighContrastColor("foreground", settings.linkLabels.fill.value.value);
         // node positions
         try {
             const nodePositionsValue = settings.nodeComplexSettings.persistProperties.nodePositions.value;
@@ -1217,7 +1253,7 @@ export class SankeyDiagram implements IVisual {
     }
 
     private adjustNodesPositions(nodes: SankeyDiagramNode[], settings: SankeyDiagramSettings) {
-        const nodeWidth: number = settings.nodesSettings.nodeWidth.value;
+        const nodeWidth: number = settings.nodes.defaultContainerItem.nodeWidth.value;
 
         // split nodes by columns
         const columns = new Map<number, SankeyDiagramNode[]>();
@@ -1400,7 +1436,7 @@ export class SankeyDiagram implements IVisual {
     }
 
     private getScaleByAxisX(numberOfColumns: number = SankeyDiagram.DefaultNumberOfColumns): number {
-        return SankeyDiagram.getPositiveNumber((this.viewport.width - this.sankeyDiagramSettings.nodesSettings.nodeWidth.value) / numberOfColumns);
+        return SankeyDiagram.getPositiveNumber((this.viewport.width - this.sankeyDiagramSettings.nodes.defaultContainerItem.nodeWidth.value) / numberOfColumns);
     }
 
     public static sortNodesByColumnIndex(nodes: SankeyDiagramNode[]): SankeyDiagramNode[] {
@@ -1727,7 +1763,7 @@ export class SankeyDiagram implements IVisual {
             .attr("x", (node: SankeyDiagramNode) => node.left - node.x)
             .attr("y", (node: SankeyDiagramNode) => node.top - node.y)
             .attr("dy", SankeyDiagram.DefaultDy)
-            .style("fill", (node: SankeyDiagramNode) => node.label.color)
+            .style("fill", (node: SankeyDiagramNode) => this.colorHelper.getHighContrastColor("foreground", node.label.color))
             .style("font-family", nodeLabelTextProperties.fontFamily)
             .style("font-size", nodeLabelTextProperties.fontSize)
             .style("font-weight", nodeLabelTextProperties.fontWeight)
@@ -1970,7 +2006,9 @@ export class SankeyDiagram implements IVisual {
             .attr("aria-selected", "false")
             .attr('aria-label', (link: SankeyDiagramLink) => `${link.source.label.name} to ${link.destination.label.name} weighted at ${link.weight}`)
             .style("stroke", (link: SankeyDiagramLink) => link.strokeColor)
-            .style("fill", (link: SankeyDiagramLink) => link.fillColor);
+            .style("stroke-width", this.sankeyDiagramSettings.links.defaultContainerItem.border.show.value ? this.sankeyDiagramSettings.links.defaultContainerItem.border.width.value + "px" : "0px")
+            .style("fill", (link: SankeyDiagramLink) => link.fillColor)
+            .classed(SankeyDiagram.StrokeVisibleClass.className, !this.sankeyDiagramSettings.links.defaultContainerItem.border.show.value && this.colorHelper.isHighContrast);
 
         return linksElements;
     }
@@ -2032,7 +2070,7 @@ export class SankeyDiagram implements IVisual {
             .style("font-weight", linkLabelsTextProperties.fontWeight)
             .style("font-style", linkLabelsTextProperties.fontStyle)
             .style("text-decoration", linkLabelsTextProperties.textDecoration)
-            .style("fill", (link: SankeyDiagramLink) => link.label.color)
+            .style("fill", (link: SankeyDiagramLink) => this.colorHelper.getHighContrastColor("foreground", link.label.color))
             .text((link: SankeyDiagramLink) => (link.label.formattedName && (link.label.formattedName.length > 0)) ? link.label.formattedName :
                 `${link.source.label.name || ""}-${link.destination.label.name || ""}:${(link.tooltipInfo[2] || { value: "" }).value}`
             );