+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.5.0.0
+
+### New features
+* Added conditional formatting to link color options
+* Added matchNodeColor setting to links
+* Added border color for links
+
 ## 3.4.6.0
 
 ### New features